refactor(web): dedupe fake SSE message bus in tests
SSEProvider.test.tsx and SSEProvider.interaction.test.tsx each hand-rolled
an identical Map<type, Set<listener>> fake mirroring sse.client's
onServerMessage. Extract one shared createFakeServerMessageBus test util.
